requirements.txt
langchain
langchain-openai
pypdf
faiss-cpu
streamlit
python-dotenv
tiktoken